Added global service messages

+++ src/Systems/ServiceMessages.h
@@ -79,8 +79,8 @@ public:
 	virtual void setPeerSerial(std::string peerSerial) { _peerSerial = peerSerial; }
 
 	virtual void load();
-	virtual void save(int64_t timestamp, uint32_t index, bool value);
-	virtual void save(int64_t timestamp, int32_t channel, std::string id, uint8_t value);
+	virtual void save(int32_t timestamp, uint32_t index, bool value);
+	virtual void save(int32_t timestamp, int32_t channel, std::string id, uint8_t value);
 	virtual bool set(std::string id, bool value);
 	virtual void set(std::string id, uint8_t value, uint32_t channel);
 	virtual std::shared_ptr<Variable> get(PRpcClientInfo clientInfo, bool returnID);
@@ -109,15 +109,15 @@ protected:
     std::string _peerSerial;
     bool _disposing = false;
     bool _configPending = false;
-    int64_t _configPendingTime = 0;
+    int32_t _configPendingTime = 0;
     int64_t _configPendingSetTime = 0;
     int32_t _unreachResendCounter = 0;
     bool _unreach = false;
-    int64_t _unreachTime = 0;
+    int32_t _unreachTime = 0;
 	bool _stickyUnreach = false;
-    int64_t _stickyUnreachTime = 0;
+    int32_t _stickyUnreachTime = 0;
 	bool _lowbat = false;
-    int64_t _lowbatTime = 0;
+    int32_t _lowbatTime = 0;
 
 	std::mutex _errorMutex;
 	std::map<uint32_t, std::map<std::string, ErrorInfo>> _errors;
